[DEV-2992] aggregatedAmountsHelper --> AwardAmountHelper
