+++ src/js/components/awardv2/idv/amounts/AggregatedAwardAmounts.jsx
@@ -7,7 +7,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { formatNumber } from 'helpers/moneyFormatter';
 
-import { determineSpendingScenario } from 'helpers/aggregatedAmountsHelper';
+import { determineSpendingScenario } from 'helpers/awardAmountHelper';
 import ChartError from 'components/search/visualizations/ChartError';
 import AwardsBanner from './AwardsBanner';
 import { AWARD_V2_AGGREGATED_AMOUNTS_PROPS } from '../../../../propTypes';
