+++ src/js/components/awardv2/shared/awardAmountsSection/charts/LoanChart.jsx
@@ -6,7 +6,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { generatePercentage } from 'helpers/aggregatedAmountsHelper';
+import { generatePercentage } from 'helpers/awardAmountHelper';
 
 import { useTooltips } from "./AwardAmountsChart";
 import TooltipWrapper from "../../../../sharedComponents/TooltipWrapper";
